+++ src/App.tsx
@@ -1,7 +1,7 @@
 import { QueryClientProvider } from "@tanstack/react-query";
 import Router from "./router/Router";
 import { queryClient } from "./api/instance";
-import { AuthProvider } from "./provider/useAuth";
+import { AuthProvider } from "./provider/AuthProvider";
 
 function App() {
   return (
