+++ src/pages/LoginModal/RedirectPage.tsx
@@ -1,6 +1,6 @@
 import { useEffect } from "react";
 import { useNavigate } from "react-router-dom";
-import useAuth from "@/provider/useAuth";
+import useAuth from "@/hooks/useAuth";
 import { apiClient } from "@/api/instance";
 import RouterPath from "@/router/RouterPath";
